+++ source/Calamari.Common/Plumbing/FileSystem/SubstituteInFiles.cs
@@ -25,7 +25,8 @@ namespace Calamari.Common.Plumbing.FileSystem
 
         public void SubstituteBasedSettingsInSuppliedVariables(string currentDirectory)
         {
-            var filesToTarget = variables.GetPaths(PackageVariables.SubstituteInFilesTargets);
+            var filesToTarget = variables.GetPaths(PackageVariables.SubstituteInFilesTargets)
+                                         .SelectMany(v => v.Split(";")).ToArray();
             Substitute(currentDirectory, filesToTarget);
         }
