+++ source/Calamari/Kubernetes/Conventions/GatherAndApplyRawYamlConvention.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -5,7 +6,6 @@ using Calamari.Common.Commands;
 using Calamari.Common.Features.Processes;
 using Calamari.Common.Plumbing.FileSystem;
 using Calamari.Common.Plumbing.Logging;
-using Calamari.Common.Plumbing.Variables;
 using Calamari.Deployment.Conventions;
 using Calamari.Kubernetes.Integration;
 
@@ -14,24 +14,22 @@ namespace Calamari.Kubernetes.Conventions;
 public class GatherAndApplyRawYamlConvention: IInstallConvention
 {
     private readonly ILog log;
-    private readonly IVariables variables;
     private readonly ICalamariFileSystem fileSystem;
     private readonly ICommandLineRunner commandLineRunner;
 
     public GatherAndApplyRawYamlConvention(
         ILog log,
-        IVariables variables,
         ICalamariFileSystem fileSystem,
         ICommandLineRunner commandLineRunner)
     {
         this.log = log;
-        this.variables = variables;
         this.fileSystem = fileSystem;
         this.commandLineRunner = commandLineRunner;
     }
 
     public void Install(RunningDeployment deployment)
     {
+        var variables = deployment.Variables;
         var rawGlobs = variables.Get("Octopus.Action.KubernetesContainers.CustomResourceYamlFileName");
         if (rawGlobs == null)
             return;
@@ -54,11 +52,21 @@ public class GatherAndApplyRawYamlConvention: IInstallConvention
         variables.Set("Octopus.Action.KubernetesContainers.YamlDirectories", string.Join(";", directories));
 
         var kubectl = new Kubectl(variables.Get("Octopus.Action.Kubernetes.CustomKubectlExecutable"), log,
-            commandLineRunner, deployment.CurrentDirectory, new Dictionary<string, string>());
+            commandLineRunner, deployment.CurrentDirectory, deployment.EnvironmentVariables);
 
-        foreach (var directory in directories)
+        if (!kubectl.TrySetKubectl())
         {
+            throw new Exception("Could not set KubeCtl");
+        }
+
+        foreach (var (directory, index) in directories.Select((d,i) => (d,i)))
+        {
+            log.Verbose($"Applying Yaml Batch #{index}");
             var output = kubectl.ExecuteCommandAndReturnOutput("apply", "-f", directory, "-o", "json");
+            foreach (var logLine in output)
+            {
+                log.Verbose(logLine);
+            }
         }
     }
 }
