+++ source/Calamari/Kubernetes/Conventions/KubernetesAuthContextConvention.cs
@@ -1,0 +1,33 @@
+using Calamari.Common.Commands;
+using Calamari.Common.Features.Processes;
+using Calamari.Common.Features.Scripts;
+using Calamari.Common.Plumbing.Logging;
+using Calamari.Deployment.Conventions;
+
+namespace Calamari.Kubernetes.Conventions;
+
+public class KubernetesAuthContextConvention : IInstallConvention
+{
+    private readonly ILog log;
+    private readonly ICommandLineRunner commandLineRunner;
+
+    public KubernetesAuthContextConvention(ILog log, ICommandLineRunner commandLineRunner)
+    {
+        this.log = log;
+        this.commandLineRunner = commandLineRunner;
+    }
+
+    public void Install(RunningDeployment deployment)
+    {
+        var setupKubectlAuthentication = new SetupKubectlAuthentication(deployment.Variables,
+            log,
+            ScriptSyntax.PowerShell,
+            commandLineRunner,
+            deployment.EnvironmentVariables,
+            deployment.CurrentDirectory);
+
+        var accountType = deployment.Variables.Get("Octopus.Account.AccountType");
+
+        setupKubectlAuthentication.Execute(accountType);
+    }
+}
